+++ pyccel/ast/class_defs.py
@@ -12,7 +12,7 @@ from pyccel.ast.builtin_methods.set_methods  import (SetAdd, SetClear, SetCopy, 
                                                      SetIsDisjoint)
 from pyccel.ast.builtin_methods.list_methods import (ListAppend, ListInsert, ListPop,
                                                      ListClear, ListExtend, ListRemove,
-                                                     ListCopy, ListSort)
+                                                     ListCopy, ListSort, ListReverse)
 from pyccel.ast.builtin_methods.dict_methods import (DictPop, DictPopitem, DictGet, DictClear,DictCopy,
                                                      DictSetDefault, DictItems, DictGetItem, DictKeys)
 
@@ -155,6 +155,7 @@ ListClass = ClassDef('list',
             PyccelFunctionDef('pop', func_class = ListPop),
             PyccelFunctionDef('sort', func_class = ListSort),
             PyccelFunctionDef('remove', func_class = ListRemove),
+            PyccelFunctionDef('reverse', func_class = ListReverse),
         ])
 
 #=======================================================================================